Move tests from BitmapTest to ContextTest class
